+++ package.json
@@ -77,6 +77,7 @@
     "react-redux": "^8.0.1",
     "react-router-dom": "^6.3.0",
     "redux-logger": "^3.0.6",
+    "redux-persist": "^6.0.0",
     "redux-thunk": "^2.4.1",
     "saas": "^1.0.0",
     "sass": "^1.51.0",
