+++ src/App.tsx
@@ -1,26 +1,15 @@
-import React, { FC, useState, Suspense } from 'react';
-import { HashRouter, Route, Routes } from 'react-router-dom';
+import React, { FC, useState } from 'react';
+import { Provider } from 'react-redux';
+import { PersistGate } from 'redux-persist/integration/react';
 
 import { ThemeContext, defaultContext } from './utils/ThemeContext';
-import { Header } from './components/Header';
-
-
-import { Home } from './pages/Home';
-import { Error } from './pages/Error';
-import { Profile } from './pages/Profile';
-import { ChatList } from './components/ChatList';
-import { AboutWithConnect } from './pages/About';
+import { AppRouter } from './components/AppRouter';
+import { store, persistor } from './store';
 import './App.css';
-
-const Chats = React.lazy(() =>
-  import('./pages/Chats').then((module) => ({
-    default: module.Chats,
-  })),
-);
 
 export const App: FC = () => {
   const [theme, setTheme] = useState(defaultContext.theme);
-  
+
   /*const chatList = useMemo(
     () =>
       Object.entries(messages).map((chat) => ({
@@ -34,40 +23,17 @@ export const App: FC = () => {
     setTheme(theme === 'light' ? 'dark' : 'light');
   };
   return (
-      <ThemeContext.Provider
-        value={{
-          theme,
-          toggleTheme,
-        }}
-      >
-        <HashRouter>
-          <Suspense fallback={<div>Loading...</div>}>
-            <Routes>
-              <Route path="/" element={<Header />}>
-                <Route index element={<Home />} />
-                <Route path="profile" element={<Profile />} />
-                <Route path="chats">
-                  <Route
-                    index
-                    element={
-                      <ChatList />
-                    }
-                  />
-                  <Route
-                    path=":chatId"
-                    element={
-                      <Chats />
-                    }
-                  />
-                </Route>
-                <Route path="about" element={<AboutWithConnect />} />
-            
-              </Route>
-
-              <Route path="*" element={<Error />} />
-            </Routes>
-          </Suspense>
-        </HashRouter>
-      </ThemeContext.Provider>
+    <ThemeContext.Provider
+      value={{
+        theme,
+        toggleTheme,
+      }}
+    >
+      <Provider store={store}>
+        <PersistGate persistor={persistor}>
+          <AppRouter />
+        </PersistGate>
+      </Provider>
+    </ThemeContext.Provider>
   );
 };
