+++ src/store/index.ts
@@ -1,6 +1,8 @@
 import { configureStore, combineReducers } from '@reduxjs/toolkit';
-import { chatReducer, ChatsState } from './chats/reducer';
-import { profileReducer, ProfileState } from './profile/reducer';
+import { chatReducer } from './chats/reducer';
+import { profileReducer } from './profile/reducer';
+import {  persistStore, persistReducer } from 'redux-persist';
+import storage from 'redux-persist/lib/storage';
 
 const rootReducer = combineReducers({
   profile: profileReducer,
@@ -9,8 +11,18 @@ const rootReducer = combineReducers({
 
 export type StoreState = ReturnType<typeof rootReducer>;
 
+const persistConfig = {
+  key: 'gb_1889',
+  storage,
+  blacklist: ['profile'],
+};
+
+const persistedReducer = persistReducer(persistConfig, rootReducer);
+
 export const store = configureStore({
-  reducer: rootReducer,
+  reducer: persistedReducer,
   devTools: process.env.NODE_ENV !== 'production',
   },
 );
+
+export const persistor =  persistStore(store);
